+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -17,13 +17,13 @@ object GradlePlugins {
 object Libs {
 
     private object Versions {
-        const val junitJupiter = "5.5.2"
+        const val junitJupiter = "5.6.0"
         const val kotlinCoroutinesRetrofit = "1.1.0"
         const val moshi = "1.9.2"
-        const val okhttp = "3.12.7"
+        const val okhttp = "3.12.8"
         const val retrofit = "2.6.4"
-        const val threetenbp = "1.4.0"
-        const val truth = "1.0"
+        const val threetenbp = "1.4.1"
+        const val truth = "1.0.1"
     }
 
     const val junitJupiterApi = "org.junit.jupiter:junit-jupiter-api:${Versions.junitJupiter}"