+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -20,7 +20,7 @@ object Libs {
         const val junitJupiter = "5.6.1"
         const val kotlinCoroutinesRetrofit = "1.1.0"
         const val moshi = "1.9.2"
-        const val okhttp = "3.12.10"
+        const val okhttp = "3.12.11"
         const val retrofit = "2.6.4"
         const val threetenbp = "1.4.3"
         const val truth = "1.0.1"