+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -22,7 +22,7 @@ object Libs {
         const val moshi = "1.9.2"
         const val okhttp = "3.12.11"
         const val retrofit = "2.6.4"
-        const val threetenbp = "1.4.3"
+        const val threetenbp = "1.4.4"
         const val truth = "1.0.1"
     }