+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -1,6 +1,6 @@
 package info.metadude.kotlin.library.engelsystem
 
-private const val kotlinVersion = "1.3.71"
+private const val kotlinVersion = "1.3.72"
 
 object GradlePlugins {
 
@@ -17,12 +17,12 @@ object GradlePlugins {
 object Libs {
 
     private object Versions {
-        const val junitJupiter = "5.6.1"
+        const val junitJupiter = "5.6.2"
         const val kotlinCoroutinesRetrofit = "1.1.0"
         const val moshi = "1.9.2"
-        const val okhttp = "3.12.10"
+        const val okhttp = "3.12.11"
         const val retrofit = "2.6.4"
-        const val threetenbp = "1.4.3"
+        const val threetenbp = "1.4.4"
         const val truth = "1.0.1"
     }