+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -19,7 +19,7 @@ object Libs {
     private object Versions {
         const val junitJupiter = "5.6.2"
         const val kotlinCoroutinesRetrofit = "1.1.0"
-        const val moshi = "1.9.2"
+        const val moshi = "1.9.3"
         const val okhttp = "3.12.12"
         const val retrofit = "2.6.4"
         const val threetenbp = "1.4.4"