+++ engelsystem-base/src/test/kotlin/info/metadude/kotlin/library/engelsystem/utils/UserAgentInterceptor.kt
@@ -5,7 +5,7 @@ import okhttp3.Interceptor.Chain
 import okhttp3.Response
 import java.io.IOException
 
-class UserAgentInterceptor(private val userAgent: String) : Interceptor {
+internal class UserAgentInterceptor(private val userAgent: String) : Interceptor {
 
     @Throws(IOException::class)
     override fun intercept(chain: Chain): Response {