+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changes
 
+* Use okhttp v.3.12.8.
 * Use threetenbp v.1.4.1.