+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 ### Changes
 
 * Use okhttp v.3.12.9.
-* Use threetenbp v.1.4.1.
+* Use threetenbp v.1.4.3.
 
 
 ## [v.3.1.0](https://github.com/johnjohndoe/engelsystem/releases/tag/v.3.1.0)