+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 
 ### Changes
 
-* Use okhttp v.3.12.9.
-* Use threetenbp v.1.4.1.
+* Use kotlin v.1.3.71.
+* Use okhttp v.3.12.10.
+* Use threetenbp v.1.4.3.
 
 
 ## [v.3.1.0](https://github.com/johnjohndoe/engelsystem/releases/tag/v.3.1.0)